feat(ssr): fix erros router and Image() used in game page
