+++ packages/client/src/core/game/game.ts
@@ -1,63 +1,8 @@
-import { IVariables } from '@/core/game/types';
+import { IGameAssets, IVariables } from '@/core/game/types';
 import { GameHooks } from '@/core/game/gameHooks';
-import { GameFactory } from '@/core/game/gameFactory';
 
-const map = [
-	'1------------21------------2',
-	'|............||............|',
-	'|.1--2.1---2.||.1---2.1--2.|',
-	'|o|  |.|   |.||.|   |.|  |o|',
-	'|.4--3.4---3.43.4---3.4--3.|',
-	'|..........................|',
-	'|.1--2.12.1------2.12.1--2.|',
-	'|.4--3.||.4--21--3.||.4--3.|',
-	'|......||....||....||......|',
-	'4----2.|4--2 || 1--3|.1----3',
-	'     |.|1--3 43 4--2|.|     ',
-	'     |.||          ||.|     ',
-	'     |.|| 1------2 ||.|     ',
-	'-----3.43 |      | 43.4-----',
-	'      .   |      |   .      ',
-	'-----2.12 |      | 12.1-----',
-	'     |.|| 4------3 ||.|     ',
-	'     |.||          ||.|     ',
-	'     |.|| 1------2 ||.|     ',
-	'1----3.43 4--21--3 43.4----2',
-	'|............||............|',
-	'|.1--2.1---2.||.1---2.1--2.|',
-	'|.4-2|.4---3.43.4---3.|1-3.|',
-	'|o..||.......  .......||..o|',
-	'4-2.||.12.1------2.12.||.1-3',
-	'1-3.43.||.4--21--3.||.43.4-2',
-	'|......||....||....||......|',
-	'|.1----34--2.||.1--34----2.|',
-	'|.4--------3.43.4--------3.|',
-	'|..........................|',
-	'4--------------------------3',
-];
-
-export const variables: IVariables = {
-	tileLength: 32,
-	isWindowVisible: true,
-	isGamePaused: false,
-	score: 0,
-	lastKeyPressed: '',
-	level: 1,
-	player: undefined, // todo:продумать что будем тут хранить всего юзера или только часть данных
-	killCount: 0,
-	start: true,
-	animationId: null,
-	directionEventListener: null,
-	visibilityEventListener: null,
-	pauseEventListener: null,
-	levelUpCount: 0,
-	frameLifetime: 10,
-	startTime: 0,
-};
-
-export const assets = GameFactory.makeAssets(map, variables);
-export function playGame(player: any) {
-	variables.animationId = requestAnimationFrame(playGame);
+export function playGame(player: any, variables: IVariables, assets: IGameAssets) {
+	variables.animationId = requestAnimationFrame(() => playGame(player, variables, assets));
 	const board = document.querySelector<HTMLCanvasElement>('#board');
 	if (board) {
 		const ctx = board!.getContext('2d');
